feat: impl triangle drawer

+++ src/Triangle.cpp
@@ -4,9 +4,9 @@
 
 #include "Triangle.h"
 
-const std::array<cv::Vec2i, 2>& Triangle::get_bounding_box() {
+const std::array<cv::Vec2f, 2>& Triangle::get_bounding_box() {
     if (data_changed) {
-        std::array bbox{cv::Vec2i(INT_MAX, INT_MAX), cv::Vec2i(INT_MIN, INT_MIN)};
+        std::array bbox{cv::Vec2f(INT_MAX, INT_MAX), cv::Vec2f(INT_MIN, INT_MIN)};
 
         for (int i = 0; i < 3; ++i) {
             bbox[0][0] = std::min(bbox[0][0], points[i][0]);
@@ -22,16 +22,16 @@ const std::array<cv::Vec2i, 2>& Triangle::get_bounding_box() {
     return bbox;
 }
 
-cv::Vec3f Triangle::barycentric(const cv::Vec2i& p) {
+cv::Vec3f Triangle::barycentric(const cv::Vec2f& p) {
     cv::Vec3f u = cv::Vec3f{
-                static_cast<float>(points[2][0] - points[0][0]),
-                static_cast<float>(points[1][0] - points[0][0]),
-                static_cast<float>(points[0][0] - p[0])
+                (points[2][0] - points[0][0]),
+                (points[1][0] - points[0][0]),
+                (points[0][0] - p[0])
             }
             .cross(cv::Vec3f{
-                static_cast<float>(points[2][1] - points[0][1]),
-                static_cast<float>(points[1][1] - points[0][1]),
-                static_cast<float>(points[0][1] - p[1])
+                (points[2][1] - points[0][1]),
+                (points[1][1] - points[0][1]),
+                (points[0][1] - p[1])
             });
 
     if (std::abs(u[2]) < 1)return cv::Vec3f{-1, 1, 1};
